+++ credits.lua
@@ -37,14 +37,17 @@ local function credits_draw()
 	Font.print(main_font, 170, 90, "CREDITS", black)
 	
 	-- reduce font size
-	Font.setPixelSizes(main_font, 20)
+	Font.setPixelSizes(main_font, 22)
 	
 	-- credit
-	Font.print(main_font, 190, 140, "Lua Player Plus Vita by Rinnegatamante (http://rinnegatamante.it)\nPeople from #henkaku irc\ntetris implementation example by codeincomplete.com\nVITA buttons by nodeadfolk\nfont : xolomium (fontlibrary.com)\nsounds : freesound.org (see github)\n\n\nBy Svennd (svennd.be)", black)
+	Font.print(main_font, 190, 140, "Lua Player Plus Vita by Rinnegatamante\n(http://rinnegatamante.it)\n\nPeople from #henkaku irc\nVITA buttons by nodeadfolk\Xolomium font (fontlibrary.com)\nsounds : freesound.org\n(see github for latest version)\n\n\n\nBy Svennd (svennd.be)", black)
 
 	-- touch tip
 	Graphics.drawImage(5, 470, img_touch, Color.new(255,255,255, 50 + math.floor(animate_touch/3)))
 
+	-- draw version
+	Graphics.drawImage(791, 506, img_version) -- version
+	
 	Graphics.termBlend()
 	Screen.flip()
 end